Smart School Management( Asp .Net Core Web API  &  WPF)
users can store both school and teacher information in the database. The application, built using ASP.NET Web API and SQL Server for the backend, with XAML for the front-end, allows users to easily create, retrieve, and manage school and teacher records. This desktop application ensures that all data is securely stored and easily accessible."